Move more unit tests to base unit tests directory
