--- test/base/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  Collection: require('./tests/collection'),
-  Events: require('./tests/events')
-};
